+++ demo/coordinate_based_network/coordinate_based_network.c
@@ -25,6 +25,149 @@
 //     }
 // }
 
+/* ------------------------------ */
+
+void nl_model_save(const char *fname, NeuralNet model)
+{
+    FILE *fptr = fopen(fname, "w");
+    if (!fptr) {
+        fprintf(stderr, "Can NOT open file '%s' to save model\n", fname);
+    }
+
+    // Write count
+    fprintf(fptr, "[Count]\n");
+    fprintf(fptr, "%zu", model.count);
+    fprintf(fptr, "\n\n");
+
+    // Write layers
+    fprintf(fptr, "[Layers]\n");
+    for (size_t i = 0; i < model.count; ++i) {
+        fprintf(fptr, "%zu ", model.layers[i]);
+    }
+    fprintf(fptr, "\n\n");
+
+    // Write activations
+    fprintf(fptr, "[Activations]\n");
+    for (size_t i = 0; i < model.count - 1; ++i) {
+        fprintf(fptr, "%u ", model.acts[i]);
+    }
+    fprintf(fptr, "\n\n");
+
+    // Write cost function type
+    fprintf(fptr, "[Cost_func]\n");
+    fprintf(fptr, "%u", model.ct);
+    fprintf(fptr, "\n\n");
+
+    // Write weights
+    fprintf(fptr, "[Weights]\n");
+    for (size_t i = 0; i < model.count - 1; ++i) {
+        fprintf(fptr, "{\n");
+        for (size_t r = 0; r < model.ws[i].rows; ++r) {
+            fprintf(fptr, "  ");
+            for (size_t c = 0; c < model.ws[i].cols; ++c) {
+                fprintf(fptr, "%f ", NL_MAT_AT(model.ws[i], r, c));
+            }
+            fprintf(fptr, "\n");
+        }
+        fprintf(fptr, "}\n");
+    }
+    fprintf(fptr, "\n");
+
+    // Write biases
+    fprintf(fptr, "[Biases]\n");
+    for (size_t i = 0; i < model.count - 1; ++i) {
+        fprintf(fptr, "{\n");
+        for (size_t r = 0; r < model.bs[i].rows; ++r) {
+            fprintf(fptr, "  ");
+            for (size_t c = 0; c < model.bs[i].cols; ++c) {
+                fprintf(fptr, "%f ", NL_MAT_AT(model.bs[i], r, c));
+            }
+            fprintf(fptr, "\n");
+        }
+        fprintf(fptr, "}\n");
+    }
+
+    fclose(fptr);
+}
+
+bool nl_model_load(const char *fname, NeuralNet *model)
+{
+    FILE *fptr = fopen(fname, "r");
+    if (!fptr) {
+        fprintf(stderr, "Can NOT open file '%s' to read\n", fname);
+    }
+
+    char buf[1024];
+    // Count
+    fscanf(fptr, "%s", buf);
+    fscanf(fptr, "%zu", &(model->count));
+    printf("%zu\n", model->count);
+
+    // Alloc memeroy
+    model->layers = NL_MALLOC(sizeof(size_t) * model->count);
+    model->acts = NL_MALLOC(sizeof(size_t) * (model->count - 1));
+    model->ws = NL_MALLOC(sizeof(Mat) * (model->count - 1));
+    model->bs = NL_MALLOC(sizeof(Mat) * (model->count - 1));
+
+    // Layers
+    fscanf(fptr, "%s", buf);
+    for (size_t i = 0; i < model->count; ++i) {
+        fscanf(fptr, "%zu ", &(model->layers[i]));
+        printf("%zu ", model->layers[i]);
+    }
+    printf("\n");
+
+    // Activations
+    fscanf(fptr, "%s", buf);
+    for (size_t i = 0; i < model->count - 1; ++i) {
+        fscanf(fptr, "%u ", &(model->acts[i]));
+        printf("%u ", model->acts[i]);
+    }
+    printf("\n");
+
+    // Cost func
+    fscanf(fptr, "%s", buf);
+    fscanf(fptr, "%u", &(model->ct));
+    printf("%s\n", buf);
+    printf("%u\n", model->ct);
+    printf("cost ok\n");
+
+    // Weights
+    fscanf(fptr, "%s", buf);
+    for (size_t i = 0; i < model->count - 1; ++i) {
+        fscanf(fptr, "%s", buf);
+        model->ws[i] = nl_mat_alloc(model->layers[i + 1], model->layers[i]);
+        for (size_t r = 0; r < model->ws[i].rows; ++r) {
+            for (size_t c = 0; c < model->ws[i].cols; ++c) {
+                fscanf(fptr, "%f ", &NL_MAT_AT(model->ws[i], r, c));
+            }
+        }
+        fscanf(fptr, "%s", buf);
+    }
+    fscanf(fptr, "%s", buf);
+    printf("weight ok\n");
+
+    // Biases
+    fscanf(fptr, "%s", buf);
+    for (size_t i = 0; i < model->count - 1; ++i) {
+        fscanf(fptr, "%s", buf);
+        model->bs[i] = nl_mat_alloc(model->layers[i + 1], 1);
+        for (size_t r = 0; r < model->bs[i].rows; ++r) {
+            for (size_t c = 0; c < model->bs[i].cols; ++c) {
+                fscanf(fptr, "%f ", &NL_MAT_AT(model->bs[i], r, c));
+            }
+        }
+        fscanf(fptr, "%s", buf);
+    }
+    fscanf(fptr, "%s", buf);
+    printf("bias ok\n");
+
+    fclose(fptr);
+}
+bool nl_model_load_with_arena(Arena *arena, const char *fname, NeuralNet *model);
+
+/* ------------------------------ */
+
 void arg_shift(int *argc, char **argv[])
 {
     if (*argc > 0) {
@@ -104,7 +247,7 @@ int main(int argc, char *argv[])
 
     printf("===== Train\n");
     float lr = 7e-2;
-    size_t epochs = 1000 * 8;
+    size_t epochs = 1000 * 1;
     size_t batch_size = 4;
     nl_model_train(model, xs, ys, lr, epochs, batch_size, false);
 
@@ -138,6 +281,13 @@ int main(int argc, char *argv[])
         }
     }
     stbi_write_png("out.png", out_w, out_h, n, img_out_data, out_w*n);
+
+    nl_model_save("model.ml", model);
+
+    NeuralNet new_model;
+    nl_model_load("model.ml", &new_model);
+    nl_model_save("test_model_read.ml", new_model);
+    nl_model_free(new_model);
 
     arena_destroy(arena);
     return 0;